+++ src/renderer/components/SelectDataset/Layout.tsx
@@ -88,15 +88,25 @@ const Layout: React.FC<{
                     Opened Datasets
                 </Typography>
                 <Carousel elementsToShow={numberOfElements}>
-                    <OpenNewCard handleOpenLocal={handleOpenLocal} />
-                    {openedFiles.reverse().map((file) => (
-                        <DatasetCard
-                            key={file.fileId}
-                            file={file}
-                            handleDatasetClose={handleDatasetClose}
-                            handleSelectFileClick={handleSelectFileClick}
-                        />
-                    ))}
+                    {[
+                        <OpenNewCard
+                            handleOpenLocal={handleOpenLocal}
+                            key="opennewcard"
+                        />,
+                    ].concat(
+                        openedFiles
+                            .reverse()
+                            .map((file) => (
+                                <DatasetCard
+                                    key={file.fileId}
+                                    file={file}
+                                    handleDatasetClose={handleDatasetClose}
+                                    handleSelectFileClick={
+                                        handleSelectFileClick
+                                    }
+                                />
+                            )),
+                    )}
                 </Carousel>
             </Box>
             <Box sx={styles.section}>